feat(presentation/placedetails): create file and refactor tests with makeSut
